CC-17 #COMMENT Add 'get person details' feature in app
A 'Get One Person' feature is added to facilitate the retrieval of an individual's information. This feature includes the addition of a search button in the 'search' component, which navigates to a new 'get-one' component displaying the selected individual's details. Modifications in models, services, and router modules are also implemented to support this addition.